Rename from_neutrader to integration
